fix: move repository interfaces to own folder
