--- src/CitMovie.Data/Repositories/ICountryRepository.cs
@@ -1,8 +1,0 @@
-namespace CitMovie.Data;
-
-public interface ICountryRepository
-{
-    Task<IEnumerable<Country>> GetAllCountriesAsync(int page, int pageSize);
-    Task<int> GetTotalCountriesCountAsync();
-
-}
