--- src/CitMovie.Data/Repositories/IGenreRepository.cs
@@ -1,7 +1,0 @@
-namespace CitMovie.Data;
-
-public interface IGenreRepository
-{
-    Task<IEnumerable<Genre>> GetAllGenresAsync(int page, int pageSize);
-    Task<int> GetTotalGenresCountAsync();
-}
