+++ src/CitMovie.Data/RepositoryInterfaces/IBookmarkRepository.cs
