+++ src/CitMovie.Data/RepositoryInterfaces/ICountryRepository.cs
@@ -1,0 +1,8 @@
+namespace CitMovie.Data;
+
+public interface ICountryRepository
+{
+    Task<IEnumerable<Country>> GetAllCountriesAsync(int page, int pageSize);
+    Task<int> GetTotalCountriesCountAsync();
+
+}
