+++ src/CitMovie.Data/RepositoryInterfaces/IGenreRepository.cs
@@ -1,0 +1,7 @@
+namespace CitMovie.Data;
+
+public interface IGenreRepository
+{
+    Task<IEnumerable<Genre>> GetAllGenresAsync(int page, int pageSize);
+    Task<int> GetTotalGenresCountAsync();
+}
