+++ src/CitMovie.Data/RepositoryInterfaces/IPromotionalMediaRepository.cs
